project1/requirements.txt
asgiref==3.5.2
Django==3.2.13
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
Pillow==9.1.1
PyJWT==2.4.0
pymongo==3.7.2
pytz==2022.1
sqlparse==0.2.4
